chore: Remove get(TrackedEntity, Program, Status) [DHIS2-17712]
