--- dhis-2/dhis-test-integration/src/test/java/org/hisp/dhis/program/EnrollmentStoreTest.java
@@ -1,183 +1,0 @@
-/*
- * Copyright (c) 2004-2022, University of Oslo
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- * Neither the name of the HISP project nor the names of its contributors may
- * be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package org.hisp.dhis.program;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import java.util.Collection;
-import java.util.Date;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-import org.hisp.dhis.common.IdentifiableObjectStore;
-import org.hisp.dhis.dbms.DbmsManager;
-import org.hisp.dhis.organisationunit.OrganisationUnit;
-import org.hisp.dhis.organisationunit.OrganisationUnitService;
-import org.hisp.dhis.program.notification.ProgramNotificationTemplate;
-import org.hisp.dhis.test.integration.PostgresIntegrationTestBase;
-import org.hisp.dhis.trackedentity.TrackedEntity;
-import org.hisp.dhis.trackedentity.TrackedEntityService;
-import org.joda.time.DateTime;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.transaction.annotation.Transactional;
-
-/**
- * @author Chau Thu Tran
- */
-@Transactional
-class EnrollmentStoreTest extends PostgresIntegrationTestBase {
-
-  @Autowired private EnrollmentStore enrollmentStore;
-
-  @Autowired private TrackedEntityService trackedEntityService;
-
-  @Autowired private OrganisationUnitService organisationUnitService;
-
-  @Autowired private ProgramService programService;
-
-  @Autowired private DbmsManager dbmsManager;
-
-  @Autowired private ProgramStageService programStageService;
-
-  @Autowired
-  @Qualifier("org.hisp.dhis.program.notification.ProgramNotificationStore")
-  private IdentifiableObjectStore<ProgramNotificationTemplate> programNotificationStore;
-
-  private Date incidentDate;
-
-  private Date enrollmentDate;
-
-  private Program programA;
-
-  private Program programB;
-
-  private Program programC;
-
-  private OrganisationUnit organisationUnitA;
-
-  private OrganisationUnit organisationUnitB;
-
-  private Enrollment enrollmentA;
-
-  private Enrollment enrollmentB;
-
-  private Enrollment enrollmentC;
-
-  private Enrollment enrollmentD;
-
-  private TrackedEntity trackedEntityA;
-
-  private Collection<Long> orgunitIds;
-
-  @BeforeEach
-  void setUp() {
-    organisationUnitA = createOrganisationUnit('A');
-    long idA = organisationUnitService.addOrganisationUnit(organisationUnitA);
-    organisationUnitB = createOrganisationUnit('B');
-    long idB = organisationUnitService.addOrganisationUnit(organisationUnitB);
-    orgunitIds = new HashSet<>();
-    orgunitIds.add(idA);
-    orgunitIds.add(idB);
-    programA = createProgram('A', new HashSet<>(), organisationUnitA);
-    programService.addProgram(programA);
-    ProgramStage stageA = new ProgramStage("StageA", programA);
-    stageA.setSortOrder(1);
-    programStageService.saveProgramStage(stageA);
-    ProgramStage stageB = new ProgramStage("StageB", programA);
-    stageB.setSortOrder(2);
-    programStageService.saveProgramStage(stageB);
-    Set<ProgramStage> programStages = new HashSet<>();
-    programStages.add(stageA);
-    programStages.add(stageB);
-    programA.setProgramStages(programStages);
-    programService.updateProgram(programA);
-    programB = createProgram('B', new HashSet<>(), organisationUnitA);
-    programService.addProgram(programB);
-    programC = createProgram('C', new HashSet<>(), organisationUnitA);
-    programService.addProgram(programC);
-    trackedEntityA = createTrackedEntity(organisationUnitA);
-    trackedEntityService.addTrackedEntity(trackedEntityA);
-    TrackedEntity trackedEntityB = createTrackedEntity(organisationUnitB);
-    trackedEntityService.addTrackedEntity(trackedEntityB);
-    DateTime testDate1 = DateTime.now();
-    testDate1.withTimeAtStartOfDay();
-    testDate1 = testDate1.minusDays(70);
-    incidentDate = testDate1.toDate();
-    DateTime testDate2 = DateTime.now();
-    testDate2.withTimeAtStartOfDay();
-    enrollmentDate = testDate2.toDate();
-    enrollmentA = new Enrollment(enrollmentDate, incidentDate, trackedEntityA, programA);
-    enrollmentA.setUid("UID-A");
-    enrollmentB = new Enrollment(enrollmentDate, incidentDate, trackedEntityA, programB);
-    enrollmentB.setUid("UID-B");
-    enrollmentB.setStatus(EnrollmentStatus.CANCELLED);
-    enrollmentC = new Enrollment(enrollmentDate, incidentDate, trackedEntityA, programC);
-    enrollmentC.setUid("UID-C");
-    enrollmentC.setStatus(EnrollmentStatus.COMPLETED);
-    enrollmentD = new Enrollment(enrollmentDate, incidentDate, trackedEntityB, programA);
-    enrollmentD.setUid("UID-D");
-  }
-
-  @Test
-  void testEventExists() {
-    enrollmentStore.save(enrollmentA);
-    enrollmentStore.save(enrollmentB);
-    dbmsManager.flushSession();
-    assertNotNull(enrollmentStore.get(enrollmentA.getId()));
-    assertNotNull(enrollmentStore.get(enrollmentB.getId()));
-    assertNull(enrollmentStore.get(99999));
-  }
-
-  @Test
-  void testGetEnrollmentsByProgram() {
-    enrollmentStore.save(enrollmentA);
-    enrollmentStore.save(enrollmentB);
-    enrollmentStore.save(enrollmentD);
-    List<Enrollment> enrollments = enrollmentStore.get(programA);
-    assertEquals(2, enrollments.size());
-    assertTrue(enrollments.contains(enrollmentA));
-    assertTrue(enrollments.contains(enrollmentD));
-    enrollments = enrollmentStore.get(programB);
-    assertEquals(1, enrollments.size());
-    assertTrue(enrollments.contains(enrollmentB));
-  }
-
-  @Test
-  void testGetExcludeDeletedEnrollment() {
-    enrollmentStore.save(enrollmentA);
-    enrollmentStore.save(enrollmentB);
-    enrollmentStore.delete(enrollmentA);
-    assertEquals(1, enrollmentStore.getAll().size());
-  }
-}
